renamed test-cases -> cases
