+++ test/cases/apply-translations.code.js
