+++ test/cases/opts-fail-on-missing.code.js
