+++ test/cases/opts-function-name.code.js
